requirements.txt
# Core dependencies for WitnessOS Divination Engines
# Phase 1: Foundation requirements

# Data validation and models
pydantic>=2.0.0

# Date/time handling
pytz>=2023.3

# Testing framework
pytest>=7.0.0

# Phase 2+ dependencies (will be uncommented as needed)

# Astronomical calculations (Phase 3)
pyswisseph>=2.10.0

# Mathematical and visualization (Phase 6)
numpy>=1.24.0
matplotlib>=3.7.0

# Image generation and graphics (Phase 6)
pillow>=10.0.0

# SVG generation (Phase 6)
svglib>=1.5.0

# API dependencies (Phase 7)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Development dependencies
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
